split tests to separate v10 sub tests
