+++ implementors/node/run-tests.ts
@@ -6,19 +6,6 @@ import { listDirectoryEntries, runFileInSubprocess } from "./tests.ts";
 const ROOT_PATH = path.resolve(import.meta.dirname, "..", "..");
 const TESTS_ROOT_PATH = path.join(ROOT_PATH, "tests");
 
-const ASSERT_MODULE_PATH = path.join(
-  ROOT_PATH,
-  "implementors",
-  "node",
-  "assert.js"
-);
-const LOAD_ADDON_MODULE_PATH = path.join(
-  ROOT_PATH,
-  "implementors",
-  "node",
-  "load-addon.js"
-);
-
 async function populateSuite(
   testContext: TestContext,
   dir: string
