+++ implementors/node/tests.ts
@@ -34,12 +34,6 @@ const GC_MODULE_PATH = path.join(
   "node",
   "gc.js"
 );
-const PROCESS_MODULE_PATH = path.join(
-  ROOT_PATH,
-  "implementors",
-  "node",
-  "process.js"
-);
 const MUST_CALL_MODULE_PATH = path.join(
   ROOT_PATH,
   "implementors",
@@ -84,8 +78,6 @@ export function runFileInSubprocess(
         "file://" + LOAD_ADDON_MODULE_PATH,
         "--import",
         "file://" + GC_MODULE_PATH,
-        "--import",
-        "file://" + PROCESS_MODULE_PATH,
         "--import",
         "file://" + MUST_CALL_MODULE_PATH,
         filePath,
