+++ tests/js-native-api/test_string/CMakeLists.txt
@@ -1,3 +1,6 @@
 add_node_api_cts_addon(test_string test_string.c test_null.c)
-# API like node_api_create_external_string_latin1 are available from NAPI_VERSION >= 10
-target_compile_definitions(test_string PRIVATE NAPI_VERSION=10)
+
+# APIs like node_api_create_external_string_latin1 and
+# node_api_create_property_key_utf8 are available from NAPI_VERSION >= 10
+add_node_api_cts_addon(test_string_v10 test_string_v10.c)
+target_compile_definitions(test_string_v10 PRIVATE NAPI_VERSION=10)
